[Synthetics] only show errors icon when error is active (#152772)
## Summary

Resolves https://github.com/elastic/kibana/issues/152760

Only shows Error Icon on the Monitor Details page when there's an active
error

### Testing

1. Create an always-up monitor
2. Navigate to the monitor details page. Ensure no icon appears next to
the Errors tab.
3. Update the monitor so that it's always down.
4. Navigate to the monitor details page. Ensure an icon appears next to
the Errors tab.

(cherry picked from commit a5d801a86c56bd8c9f855220e4dbd3bd5b50b915)
